+++ TaiChi/moveTaiChi.h
@@ -48,29 +48,29 @@ class Move
 {
 public:
     Move();
-    Move(double global_speed_rate);
+    Move(float global_speed_rate);
 
-    void SetGlobalSpeedRate(double global_speed_rate); //设置全局速度比率
+    void SetGlobalSpeedRate(float global_speed_rate); //设置全局速度比率
 
     uint8_t GetCurrentMove(void); //获取当前运动方向
-    double GetCurrentSpeedRate(void); //获取当前运动速度比率
+    float GetCurrentSpeedRate(void); //获取当前运动速度比率
 
-    void Wheel(uint8_t wheel, uint8_t rotation, double speed_rate = 1.0); //控制某个轮子转动
+    void Wheel(uint8_t wheel, uint8_t rotation, float speed_rate = 1.0); //控制某个轮子转动
 
-    void Forward(double speed_rate = 1.0); //前进
-    void Backward(double speed_rate = 1.0); //后退
-    void ForLeftward(double speed_rate = 1.0, double turn_speed_rate = DEFAULT_TRUN_SPEED_RATE); //向前左转
-    void ForRightward(double speed_rate = 1.0, double turn_speed_rate = DEFAULT_TRUN_SPEED_RATE); //向前右转
-    void BackLeftward(double speed_rate = 1.0, double turn_speed_rate = DEFAULT_TRUN_SPEED_RATE); //向后左转
-    void BackRightward(double speed_rate = 1.0, double turn_speed_rate = DEFAULT_TRUN_SPEED_RATE); //向后右转
+    void Forward(float speed_rate = 1.0); //前进
+    void Backward(float speed_rate = 1.0); //后退
+    void ForLeftward(float speed_rate = 1.0, float turn_speed_rate = DEFAULT_TRUN_SPEED_RATE); //向前左转
+    void ForRightward(float speed_rate = 1.0, float turn_speed_rate = DEFAULT_TRUN_SPEED_RATE); //向前右转
+    void BackLeftward(float speed_rate = 1.0, float turn_speed_rate = DEFAULT_TRUN_SPEED_RATE); //向后左转
+    void BackRightward(float speed_rate = 1.0, float turn_speed_rate = DEFAULT_TRUN_SPEED_RATE); //向后右转
     void Stop(void); //制动
 
-    void MoveDirection(uint8_t direction, double speed_rate = 1.0, double turn_speed_rate = DEFAULT_TRUN_SPEED_RATE); //向某方向运动
+    void MoveDirection(uint8_t direction, float speed_rate = 1.0, float turn_speed_rate = DEFAULT_TRUN_SPEED_RATE); //向某方向运动
 
 private:
-    double global_speed_rate; //全局速度比率
+    float global_speed_rate; //全局速度比率
     uint8_t current_move; //当前运动状态
-    double current_speed_rate; //当前运动速度比率
+    float current_speed_rate; //当前运动速度比率
 };
